pyproject.toml
[project]
name = "simple-torch-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0.0",
    "graphviz>=0.20.3",
]
